% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{aggregate_multiscale_data}
\alias{aggregate_multiscale_data}
\title{Aggregate Multiscale Data to a Target Resolution}
\usage{
aggregate_multiscale_data(data, levels, hierarchy, target_lvl)
}
\arguments{
\item{data}{List of dataframes}

\item{levels}{List mapping variables to levels}

\item{hierarchy}{Hierarchy string (e.g. "Year > Individual")}

\item{target_lvl}{The resolution to aggregate everything to}
}
\value{
List of dataframes truncated/aggregated to stop at target_lvl
}
\description{
Collapses finer-level variables to a coarser grain by taking the mean.
Used for cross-scale d-separation tests where tests must be locked to the
resolution of the coarsest variable involved.
}
\keyword{internal}
