% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{auto_detect_hierarchical}
\alias{auto_detect_hierarchical}
\title{Auto-Detect Multiscale Data Structure}
\usage{
auto_detect_hierarchical(data, eq_vars, quiet = FALSE)
}
\arguments{
\item{data}{List of data.frames at different scales}

\item{eq_vars}{Character vector of variable names used in equations}

\item{quiet}{Logical, if TRUE suppress messages}
}
\value{
List with 'levels', 'hierarchy', 'link_vars'
}
\description{
Infers scales, hierarchy, and link variables from a list of data.frames
}
\keyword{internal}
