% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{auto_stack_multispecies_data}
\alias{auto_stack_multispecies_data}
\title{Auto-Stack Multispecies Data}
\usage{
auto_stack_multispecies_data(data, equations, quiet = FALSE)
}
\arguments{
\item{data}{Input list of data (e.g. list(Y=list(Sp1=mat...), Hab=vec, Trait=vec))}

\item{equations}{List of model equations}

\item{quiet}{Logical, suppress messages}
}
\value{
List with components:
\item{data}{New stacked dataframe}
\item{random_part}{String to append to random formula (e.g. "+ (1|SpeciesID)")}
\item{is_stacked}{Logical, whether stacking occurred}
}
\description{
Converts a list of matrices (Wide) into a stacked dataframe (Long) for multiscale analysis.
Replicates site and species covariates accordingly.
}
\keyword{internal}
