% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mag_helpers.R
\name{dag_matrix_to_dagitty}
\alias{dag_matrix_to_dagitty}
\title{Convert an adjacency matrix to a dagitty string}
\usage{
dag_matrix_to_dagitty(dag, latents = NULL)
}
\arguments{
\item{dag}{Named adjacency matrix (1 for edge, 0 for no edge)}

\item{latents}{Optional character vector of latent variable names}
}
\value{
A character string in dagitty format
}
\description{
Convert an adjacency matrix to a dagitty string
}
\keyword{internal}
