% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_equations_hook.default}
\alias{dsep_equations_hook.default}
\title{Default Method for D-Sep Equations Hook}
\usage{
\method{dsep_equations_hook}{default}(family, equations, dsep_equations, ...)
}
\description{
Default Method for D-Sep Equations Hook
}
\keyword{internal}
