% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_equations_hook}
\alias{dsep_equations_hook}
\title{Expanded D-Sep Equation Hook}
\usage{
dsep_equations_hook(family, equations, dsep_equations, ...)
}
\arguments{
\item{family}{The S3 family list.}

\item{equations}{The full list of model equations.}

\item{dsep_equations}{The current list of equations for the test.}

\item{...}{Additional arguments.}
}
\value{
A modified list of equations.
}
\description{
Modules implement this to add supporting equations (e.g. detection models)
to a d-separation test.
}
\keyword{internal}
