% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_potential_latent_hook.default}
\alias{dsep_potential_latent_hook.default}
\title{Default Method for D-Sep Potential Latents}
\usage{
\method{dsep_potential_latent_hook}{default}(family, potential_latents, ...)
}
\description{
Default Method for D-Sep Potential Latents
}
\keyword{internal}
