% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_potential_latent_hook}
\alias{dsep_potential_latent_hook}
\title{D-Sep Potential Latent Hook}
\usage{
dsep_potential_latent_hook(family, potential_latents, ...)
}
\arguments{
\item{family}{The S3 family list.}

\item{potential_latents}{Character vector of variables suspected to be latent.}

\item{...}{Additional arguments.}
}
\value{
A modified character vector of potential latents.
}
\description{
Modules implement this to remove variables from the 'potential latents' list (e.g. occupancy states).
}
\keyword{internal}
