% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_test_translation_hook.default}
\alias{dsep_test_translation_hook.default}
\title{Default Method for D-Sep Translation}
\usage{
\method{dsep_test_translation_hook}{default}(family, test_eq, ...)
}
\description{
Default Method for D-Sep Translation
}
\keyword{internal}
