% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_test_translation_hook}
\alias{dsep_test_translation_hook}
\title{D-Sep Test Translation Hook}
\usage{
dsep_test_translation_hook(family, test_eq, ...)
}
\arguments{
\item{family}{The S3 family list.}

\item{test_eq}{The d-separation test equation.}

\item{...}{Additional arguments.}
}
\value{
A modified test equation.
}
\description{
Modules implement this to translate test equations (e.g. psi_Species ~ X to Species ~ X).
}
\keyword{internal}
