% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_tree_hook}
\alias{dsep_tree_hook}
\title{D-Sep Tree Hook}
\usage{
dsep_tree_hook(tree, test_eq, hierarchical_info, levels, ...)
}
\arguments{
\item{tree}{The structure object (e.g. phylo).}

\item{test_eq}{The current d-separation test equation.}

\item{hierarchical_info}{Hierarchical metadata.}

\item{levels}{Level mapping.}

\item{...}{Additional arguments.}
}
\value{
The tree object or NULL.
}
\description{
Modules implement this to decide if a tree should be passed to a specific d-separation test.
}
\keyword{internal}
