% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_definitions.R
\docType{data}
\name{family_definitions}
\alias{family_definitions}
\alias{.nimble_fn_cache}
\title{Family Definition Generics and Default Methods}
\format{
An object of class \code{environment} of length 0.
}
\usage{
.nimble_fn_cache
}
\description{
S3 generics and methods for distribution families in because.
This enables custom distributions to be added by defining S3 methods.
}
\keyword{datasets}
