% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deterministic_nodes.R
\name{generate_deterministic_jags}
\alias{generate_deterministic_jags}
\title{Generate JAGS Code for Deterministic Nodes}
\usage{
generate_deterministic_jags(deterministic_terms)
}
\arguments{
\item{deterministic_terms}{List from extract_deterministic_terms}
}
\value{
Vector of JAGS code strings
}
\description{
writers the "Internal_Var\[i] <- ..." lines
}
\keyword{internal}
