% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{get_data_for_variables}
\alias{get_data_for_variables}
\title{Get Data for Variables}
\usage{
get_data_for_variables(
  variables,
  data,
  levels,
  hierarchy,
  link_vars,
  equations = NULL,
  latent = NULL
)
}
\arguments{
\item{variables}{Character vector of variable names}

\item{data}{List of data.frames at different levels}

\item{levels}{List mapping variable names to level names}

\item{hierarchy}{Character string specifying nesting}

\item{link_vars}{Character vector of linking variables}
}
\value{
data.frame
}
\description{
Determine the finest grain level needed for a set of variables
and return the appropriate dataset (with joining if needed)
}
\keyword{internal}
