% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_occupancy.R
\name{get_family_object}
\alias{get_family_object}
\title{Get Family Object for S3 Dispatch}
\usage{
get_family_object(family_name)
}
\arguments{
\item{family_name}{The name of the family (e.g., "occupancy", "gaussian").}
}
\value{
An object of class \verb{because_family_<name>} and \code{because_family}.
}
\description{
Converts a family name string into a family class object for S3 dispatch.
This provides a general mechanism for family extensions to be used.
}
\keyword{internal}
