% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/because_dsep.R
\name{get_inherited_random_terms}
\alias{get_inherited_random_terms}
\title{Get inherited random terms for a variable based on hierarchy}
\usage{
get_inherited_random_terms(var, hierarchical_info)
}
\arguments{
\item{var}{Variable name}

\item{hierarchical_info}{List with $levels, $hierarchy, and $link_vars}
}
\value{
List of random terms (group, type)
}
\description{
Get inherited random terms for a variable based on hierarchy
}
\keyword{internal}
