% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_inits_hook.default}
\alias{get_inits_hook.default}
\title{Default Method for Initialization Hook}
\usage{
\method{get_inits_hook}{default}(family, data, ...)
}
\description{
Default Method for Initialization Hook
}
\keyword{internal}
