% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_inits_hook}
\alias{get_inits_hook}
\title{Initialization Hook}
\usage{
get_inits_hook(family, data, ...)
}
\arguments{
\item{family}{The S3 family list.}

\item{data}{The model data.}

\item{...}{Additional arguments.}
}
\value{
A named list of initial values.
}
\description{
Modules implement this to provide specialized initial values (e.g. latent states for occupancy).
}
\keyword{internal}
