% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_monitor_vars_hook.default}
\alias{get_monitor_vars_hook.default}
\title{Default Method for Monitor Variables Hook}
\usage{
\method{get_monitor_vars_hook}{default}(family, variable_name, ...)
}
\description{
Default Method for Monitor Variables Hook
}
\keyword{internal}
