% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_order_labels_hook.default}
\alias{get_order_labels_hook.default}
\title{Default Method for Order Labels Hook}
\usage{
\method{get_order_labels_hook}{default}(structure, ...)
}
\description{
Default Method for Order Labels Hook
}
\keyword{internal}
