% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_structure_name_hook.default}
\alias{get_structure_name_hook.default}
\title{Default Method for Structure Name}
\usage{
\method{get_structure_name_hook}{default}(structure, ...)
}
\description{
Default Method for Structure Name
}
\keyword{internal}
