% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_structure_name_hook}
\alias{get_structure_name_hook}
\title{Structure Name Hook}
\usage{
get_structure_name_hook(structure, ...)
}
\arguments{
\item{structure}{The structural object.}

\item{...}{Additional arguments.}
}
\value{
Character string.
}
\description{
Modules implement this to provide specialized names for structure precision matrices.
}
\keyword{internal}
