% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_tree_hook.default}
\alias{get_tree_hook.default}
\title{Default Method for Tree Extraction Hook}
\usage{
\method{get_tree_hook}{default}(structure, ...)
}
\description{
Default Method for Tree Extraction Hook
}
\keyword{internal}
