% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_tree_hook}
\alias{get_tree_hook}
\title{Tree Extraction Hook}
\usage{
get_tree_hook(structure, ...)
}
\arguments{
\item{structure}{The structural object.}

\item{...}{Additional arguments.}
}
\value{
A structure object (e.g. phylo) or NULL.
}
\description{
Modules implement this to extract the appropriate tree/structure (e.g. tip extraction from multiPhylo).
}
\keyword{internal}
