% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_variability_type_hook.default}
\alias{get_variability_type_hook.default}
\title{Default Method for Variability Type Hook}
\usage{
\method{get_variability_type_hook}{default}(family, variable_name, ...)
}
\description{
Default Method for Variability Type Hook
}
\keyword{internal}
