% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_latent_child_hook.default}
\alias{is_latent_child_hook.default}
\title{Default Method for Latent Child Hook}
\usage{
\method{is_latent_child_hook}{default}(family, variable_name, ...)
}
\description{
Default Method for Latent Child Hook
}
\keyword{internal}
