% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{is_latent_child_hook}
\alias{is_latent_child_hook}
\title{Latent Child Hook}
\usage{
is_latent_child_hook(family, variable_name, ...)
}
\arguments{
\item{family}{The S3 family list.}

\item{variable_name}{Name of the variable.}

\item{...}{Additional arguments.}
}
\value{
Logical.
}
\description{
Modules implement this if a variable is modeled via a latent state (e.g. occupancy Y -> z_Y).
}
\keyword{internal}
