% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jags_family_definition.default}
\alias{jags_family_definition.default}
\title{Default Method for JAGS Family Definition}
\usage{
\method{jags_family_definition}{default}(family, response, predictors, ...)
}
\description{
Default Method for JAGS Family Definition
}
\keyword{internal}
