% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_definitions.R
\name{jags_family_precision_prior}
\alias{jags_family_precision_prior}
\title{Get default precision prior for a distribution family}
\usage{
jags_family_precision_prior(family, param_name, ...)
}
\arguments{
\item{family}{A family object}

\item{param_name}{Name of the parameter (e.g. "tau_e_Abundance")}

\item{...}{Additional arguments}
}
\value{
Character string; JAGS prior statement (e.g. "tau_e_Abundance ~ dgamma(1, 1)")
}
\description{
Get default precision prior for a distribution family
}
