% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{jags_structure_definition.matrix}
\alias{jags_structure_definition.matrix}
\title{Method for JAGS Structure Definition (Matrix)}
\usage{
\method{jags_structure_definition}{matrix}(structure, variable_name = "err", ...)
}
\description{
Method for JAGS Structure Definition (Matrix)
}
