% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{needs_zero_inflation_hook.default}
\alias{needs_zero_inflation_hook.default}
\title{Default Method for Zero Inflation Hook}
\usage{
\method{needs_zero_inflation_hook}{default}(family, variable_name, ...)
}
\description{
Default Method for Zero Inflation Hook
}
\keyword{internal}
