% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{needs_zero_inflation_hook}
\alias{needs_zero_inflation_hook}
\title{Needs Zero Inflation Hook}
\usage{
needs_zero_inflation_hook(family, variable_name, ...)
}
\arguments{
\item{family}{The S3 family list.}

\item{variable_name}{Name of the response variable.}

\item{...}{Additional arguments.}
}
\value{
Logical.
}
\description{
Modules implement this to signal if a variable needs a zero-inflated likelihood.
}
\keyword{internal}
