% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_definitions.R
\name{nimble_distributions}
\alias{nimble_distributions}
\alias{dnb_because}
\alias{rnb_because}
\alias{dzip_because}
\alias{rzip_because}
\alias{dzinb_because}
\alias{rzinb_because}
\title{Custom NIMBLE Distributions for because}
\usage{
dnb_because(...)

rnb_because(...)

dzip_because(...)

rzip_because(...)

dzinb_because(...)

rzinb_because(...)
}
\arguments{
\item{...}{Arguments passed to the distribution.}
}
\description{
These functions provide custom distribution implementations for NIMBLE.
}
\keyword{internal}
