% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nimble_family_optimization.default}
\alias{nimble_family_optimization.default}
\title{Default Method for NIMBLE Optimization}
\usage{
\method{nimble_family_optimization}{default}(family, model_string, ...)
}
\description{
Default Method for NIMBLE Optimization
}
