% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{nimble_family_optimization}
\alias{nimble_family_optimization}
\title{NIMBLE Family Optimization}
\usage{
nimble_family_optimization(family, model_string, ...)
}
\arguments{
\item{family}{The S3 family object.}

\item{model_string}{The current model code string.}

\item{...}{Additional arguments.}
}
\value{
A list with \code{model_string} and \code{nimble_functions} (list of nimbleFunction).
}
\description{
Modules implement this to provide specialized NIMBLE optimizations (e.g. marginalization).
}
