% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{normalize_equations_hook.default}
\alias{normalize_equations_hook.default}
\title{Default Method for Normalize Equations Hook}
\usage{
\method{normalize_equations_hook}{default}(family, equations, ...)
}
\description{
Default Method for Normalize Equations Hook
}
\keyword{internal}
