% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{normalize_equations_hook}
\alias{normalize_equations_hook}
\title{Normalize Equations Hook}
\usage{
normalize_equations_hook(family, equations, ...)
}
\arguments{
\item{family}{The S3 family list.}

\item{equations}{The list of structural formulas.}

\item{...}{Additional arguments.}
}
\value{
A modified list of formulas.
}
\description{
Modules implement this to normalize specialized aliases in formulas (e.g. psi_Species -> Species).
}
\keyword{internal}
