% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{numpyro_structure_definition.default}
\alias{numpyro_structure_definition.default}
\title{Default Method for NumPyro Structure Definition}
\usage{
\method{numpyro_structure_definition}{default}(structure, variable_name = "err", ...)
}
\description{
Default Method for NumPyro Structure Definition
}
\keyword{internal}
