% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{numpyro_structure_definition}
\alias{numpyro_structure_definition}
\title{NumPyro Structure Definition}
\usage{
numpyro_structure_definition(structure, variable_name = "err", ...)
}
\arguments{
\item{structure}{The structural object.}

\item{variable_name}{The name of the variable.}

\item{...}{Additional arguments.}
}
\value{
A character string of valid JAX Python code.
}
\description{
Defines the Python code (JAX logic) to inject for a custom structure when using
\code{engine = "numpyro"}. Extension packages (e.g. \code{because.phybase})
should implement this S3 generic to support their structures natively in Python.
}
\keyword{internal}
