% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{parse_hierarchy_from_random}
\alias{parse_hierarchy_from_random}
\title{Parse Hierarchy from Random Effects}
\usage{
parse_hierarchy_from_random(random, data = NULL)
}
\arguments{
\item{random}{Formula specifying random effects}

\item{data}{List of data.frames at different hierarchical levels (optional, currently unused)}
}
\value{
Character string hierarchy (e.g., "site_year > individual") or NULL
}
\description{
Extract multiscale nesting structure from random effects formula
}
\keyword{internal}
