% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_coef}
\alias{plot_coef}
\title{Plot Path Coefficients}
\usage{
plot_coef(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments.}
}
\description{
Plot Path Coefficients
}
