% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Posterior Predictive Samples}
\usage{
posterior_predict(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments.}
}
\value{
A matrix of posterior predictive draws.
}
\description{
Posterior Predictive Samples
}
