% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_prep.R
\name{prepare_hierarchical_jags_data}
\alias{prepare_hierarchical_jags_data}
\title{Prepare Multiscale Data for JAGS}
\usage{
prepare_hierarchical_jags_data(hierarchical_info, vars_needed)
}
\arguments{
\item{hierarchical_info}{List containing 'data', 'levels', 'hierarchy', 'link_vars'}

\item{vars_needed}{Character vector of all variable names needed for the model}
}
\value{
List with 'data_list' (for jags) and 'n_vec' (named vector of sample sizes)
}
\description{
Transforms multiscale dataframes into a flat list of vectors and ID indices suitable for JAGS
}
\keyword{internal}
