% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prepare_structure_data.default}
\alias{prepare_structure_data.default}
\title{Default Method for Structure Data Preparation}
\usage{
\method{prepare_structure_data}{default}(structure, data, ...)
}
\description{
Default Method for Structure Data Preparation
}
\keyword{internal}
