% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prepare_structure_data.matrix}
\alias{prepare_structure_data.matrix}
\title{Default Method for Structure Definition (Independent Model)}
\usage{
\method{prepare_structure_data}{matrix}(structure, data, optimize = TRUE, ...)
}
\description{
Default Method for Structure Definition (Independent Model)
}
