% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prepare_structure_data}
\alias{prepare_structure_data}
\title{Prepare Structure Data}
\usage{
prepare_structure_data(structure, data, ...)
}
\arguments{
\item{structure}{The structure object.}

\item{data}{The model data.}

\item{...}{Additional arguments.}
}
\value{
A named list of data to be passed to JAGS (e.g., list(VCV = ...)).
}
\description{
Modules implement this to process the structure object into data for JAGS.
}
\keyword{internal}
