% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/because.R
\name{preprocess_categorical_vars}
\alias{preprocess_categorical_vars}
\title{Preprocess categorical variables (character/factor) to integer codes and dummies}
\usage{
preprocess_categorical_vars(
  data,
  target_vars = NULL,
  dummy_vars = NULL,
  exclude_cols = NULL,
  quiet = FALSE,
  expand_ordered = FALSE
)
}
\arguments{
\item{data}{A data.frame or list of data.frames}

\item{quiet}{Logical; whether to suppress informational messages}
}
\value{
The modified data object with categorical_vars attribute
}
\description{
Preprocess categorical variables (character/factor) to integer codes and dummies
}
\keyword{internal}
