% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.because.R
\name{print.summary.because}
\alias{print.summary.because}
\title{Print Summary for Because Model}
\usage{
\method{print}{summary.because}(x, ...)
}
\arguments{
\item{x}{A summary object of class \code{"summary.because"}.}

\item{...}{Additional arguments.}
}
\description{
Print Summary for Because Model
}
\keyword{internal}
