% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.because.R
\name{summary.because}
\alias{summary.because}
\title{Summary for Because Model}
\usage{
\method{summary}{because}(
  object,
  show_internal = FALSE,
  show_nodes = FALSE,
  show_random = FALSE,
  ...
)
}
\arguments{
\item{object}{A \code{because} object.}

\item{show_internal}{Logical. If \code{TRUE}, shows internal parameters created for
deterministically defined nodes (e.g., \code{beta_..._det_...}). Defaults to \code{FALSE}.}

\item{show_nodes}{Logical. If \code{TRUE}, shows latent node values (e.g., \code{Age[1]}).
Defaults to \code{FALSE} to prevent clutter when \code{monitor="all"}.}

\item{show_random}{Logical. If \code{TRUE}, shows random effect estimates (e.g., \code{u_...}).
Defaults to \code{FALSE}.}

\item{...}{Additional arguments passed to \code{\link[coda]{summary.mcmc}}.}
}
\value{
A summary object containing statistics for the monitored parameters.
If \code{dsep = TRUE} was used in \code{because}, the summary focuses on
the conditional independence tests.
}
\description{
Summarizes the output of a Because model run.
}
\examples{
\dontrun{
fit <- because(list(Y ~ X), data = my_data)

# Standard summary
summary(fit)

# Show latent states and random effects
summary(fit, show_nodes = TRUE, show_random = TRUE)
}

}
