% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/because_do.R
\name{summary.because_counterfactual}
\alias{summary.because_counterfactual}
\title{Summarize a Because Counterfactual Object}
\usage{
\method{summary}{because_counterfactual}(object, ...)
}
\arguments{
\item{object}{An object of class \code{because_counterfactual} returned by \code{do()}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A data.frame summarizing the expected value and credible intervals for each variable.
}
\description{
Summarize a Because Counterfactual Object
}
