% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{transform_graph_for_dsep}
\alias{transform_graph_for_dsep}
\title{Transform Graph for D-Separation}
\usage{
transform_graph_for_dsep(family, equations, ...)
}
\arguments{
\item{family}{The S3 family object.}

\item{equations}{The list of structural formulas.}

\item{...}{Additional arguments.}
}
\value{
The transformed list of formulas.
}
\description{
Modules implement this to rename variables in the causal graph to match their latent structure.
}
\keyword{internal}
