% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_data.R
\name{validate_hierarchical_data}
\alias{validate_hierarchical_data}
\title{Validate Hierarchical Data Structure}
\usage{
validate_hierarchical_data(
  data,
  levels,
  hierarchy,
  link_vars,
  latent_vars = NULL
)
}
\arguments{
\item{data}{List of data.frames at different hierarchical levels}

\item{levels}{List mapping variable names to level names}

\item{hierarchy}{Character string specifying nesting (e.g., "site_year > individual")}

\item{link_vars}{Character vector of variables that link levels}
}
\description{
Validate Hierarchical Data Structure
}
\keyword{internal}
