% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{dsep_tree_hook.default}
\alias{dsep_tree_hook.default}
\title{Default Method for D-Sep Tree Hook}
\usage{
\method{dsep_tree_hook}{default}(tree, test_eq, hierarchical_info, levels, ...)
}
\description{
Default Method for D-Sep Tree Hook
}
\keyword{internal}
