% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{plot_dsep}
\alias{plot_dsep}
\title{Plot D-Separation Tests}
\usage{
plot_dsep(object, ...)
}
\arguments{
\item{object}{A \code{because} object fitted with \code{dsep = TRUE}.}

\item{...}{Additional arguments.}
}
\description{
Visualizes the results of d-separation tests as a caterpillar plot.
}
