% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rope.R
\name{plot_rope}
\alias{plot_rope}
\title{Plot Region of Practical Equivalence (ROPE)}
\usage{
plot_rope(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments.}
}
\description{
Generic for plotting posterior density distributions and visualizing their overlap
with a Region of Practical Equivalence (ROPE).
}
