% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{posterior_predict}
\alias{posterior_predict}
\title{Posterior Predictive Samples}
\usage{
posterior_predict(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A matrix of posterior predictive draws.
}
\description{
Generic function for generating posterior predictive draws from a fitted model.
For \code{because} model objects, see \code{\link{posterior_predict.because}},
which returns an \code{[ndraws x N_obs]} matrix of simulated response values.
These draws are the basis for \code{\link{pp_check}}.
}
