% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pp_check}
\alias{pp_check}
\title{Posterior Predictive Checks}
\usage{
pp_check(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to the method.}
}
\value{
A ggplot object.
}
\description{
Generic function for posterior predictive checks of a fitted model.
For \code{because} model objects, see \code{\link{pp_check.because}},
which wraps \code{bayesplot} functions (density overlay, histogram, test
statistics) and supports conditional or marginal prediction via
\code{re_formula}.
}
