% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{pp_check}
\alias{pp_check}
\title{Posterior Predictive Checks}
\usage{
pp_check(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments.}
}
\value{
A ggplot object.
}
\description{
Posterior Predictive Checks
}
