% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{prepare_structure_data.matrix}
\alias{prepare_structure_data.matrix}
\title{Method for Structure Data Preparation (Matrix)}
\usage{
\method{prepare_structure_data}{matrix}(structure, data, ...)
}
\description{
Method for Structure Data Preparation (Matrix)
}
