% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{transform_graph_for_dsep.default}
\alias{transform_graph_for_dsep.default}
\title{Default Method for D-Sep Transformation}
\usage{
\method{transform_graph_for_dsep}{default}(family, equations, ...)
}
\description{
Default Method for D-Sep Transformation
}
\keyword{internal}
